cleaned up track slice
