+++ spotify-ui/src/constants/types/common.ts
@@ -1,0 +1,3 @@
+export type ExternalUrls = { [key: string]: string };
+
+export type AvailableMarkets = Array<string>;
