+++ spotify-ui/src/constants/types/index.ts
@@ -1,0 +1,1 @@
+export default {}
