+++ spotify-ui/src/store/slices/track/types.ts
@@ -1,65 +1,10 @@
-import { Artists } from '../../../components/playback/context/types'
-
-type ExternalUrls = { [key: string]: string };
-type AvailableMarkets = Array<string>;
-interface Artist extends Artists {
-  type: string;
-  id: string;
-  href: string;
-  external_urls: ExternalUrls;
-}
-
-interface Image {
-  height: number;
-  url: string;
-  width: number;
-}
-
-type Album = {
-  album_type: string;
-  artists: Array<Artist>;
-  available_markets: AvailableMarkets;
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  images: Array<Image>
-  name: string;
-  release_date: string;
-  release_date_precision: string;
-  total_tracks: number;
-  type: string;
-  uri: string;
-}
-
-type TrackDetails = {
-  album: Album;
-  artists: Array<Artist>;
-  available_markets: AvailableMarkets;
-  disc_number: number;
-  duration_ms: number;
-  explicit: boolean;
-  externalIds: ExternalUrls;
-  externalUrls: ExternalUrls;
-  href: string;
-  id: string;
-  is_local: boolean;
-  name: string;
-  popularity: number;
-  preview_url: string;
-  track_number: number;
-  type: string;
-  uri: string;
-}
-export type Track = {
-  added_at: string;
-  track: TrackDetails;
-}
+import { Track } from "../../../constants/types/track";
 
 export type SavedTracks = {
   tracks: Array<Track>;
-  offset?: number; // index of the next request start point
-  total?: number;
-  limit?: number;
+  offset: number; // * index of the next request start point
+  total: number;
+  limit: number; // * amount per request
 }
 export type TrackState = {
   savedTracks: SavedTracks;
